api_project/requirements.txt
Django==1.10
Pillow==3.3.0
requests==2.11.0
djangorestframework==3.4.4
django-filter==0.14.0
Markdown==2.6.6
